pipeline/requirements.txt
anthropic
google-cloud-storage
litellm
pdf2image
Pillow
psycopg2-binary
python-dotenv
tqdm
openpyxl
pdfplumber
requests
beautifulsoup4
